dev dump

+++ src/csharp/common/ConnectionManager.cs
@@ -11,7 +11,7 @@ namespace Mas.Infrastructure.Common
         private ConcurrentDictionary<string, TcpRpcClient> _Connections = new();
         private TcpRpcServer _Server;
 
-        public int Port { get { return _Server.Port; } }
+        public ushort Port { get { return (ushort)_Server.Port; } }
 
         public ConnectionManager() 
         {
